feat: refactor dashboard structure to module
