+++ src/modules/dashboard/dashboard.route.ts
@@ -1,0 +1,5 @@
+const dashboardRoutes = {
+    dashboard: "/dashboard",
+} as const;
+
+export default dashboardRoutes;
